requirements-dev.txt
pre-commit==3.5.0
ruff==0.15.0
pytest==8.3.5
pytest-asyncio==0.24.0
pyjwt==2.7.0
requests
imageio
imageio-ffmpeg
jsonschema
